modular ante with tx.gas
