+++ types/store.go
@@ -128,3 +128,14 @@ type (
 func NewInfiniteGasMeter() GasMeter {
 	return types.NewInfiniteGasMeter()
 }
+
+// SetGasMeter returns a new context with a gas meter set from a given context.
+func SetGasMeter(simulate bool, ctx sdk.Context, gasLimit uint64) sdk.Context {
+	// In various cases such as simulation and during the genesis block, we do not
+	// meter any gas utilization.
+	if simulate || ctx.BlockHeight() == 0 {
+		return ctx.WithGasMeter(sdk.NewInfiniteGasMeter())
+	}
+
+	return ctx.WithGasMeter(sdk.NewGasMeter(gasLimit))
+}
